% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carousel.R
\name{carousel}
\alias{carousel}
\title{Make html tags for embeding Bootstrap carousel}
\usage{
carousel(id, img, caption_title = NULL, caption_text, interval = "5000",
  pause = "hover", wrap = TRUE, keyboard = TRUE)
}
\arguments{
\item{id}{character. html identifiler.}

\item{img}{characters vector. Path to imgs.}

\item{caption_title}{charactors vector. Titles of image caption. This should have the same length of img.}

\item{caption_text}{charactors vector. Texts of image caption. This should have the same length of img.}

\item{interval}{integer. The amount of time to delay between automatically cycleing an item. If false, carousel will not automatically cycle.}

\item{pause}{string. If set to "hover", pauses the cycling of the carousel on mouseenter and resumes the cycling of the carousel on mouseleave. If set to null, hovering over the carousel won't pause it.}

\item{wrap}{Logical. Whether the carousel should cycle continuously or have hard stops.}

\item{keyboad}{Logical. Whether the carousel should react to keyboard events.}
}
\description{
Make html tags for embeding Bootstrap carousel
}
\examples{

# set imgs
png("plot1.png")
plot(1:10)
dev.off()
png("plot2.png")
plot(10:1)
dev.off()

# insert the function into a R chunk
carousel(id = "carousel_id", img = c("plot1.png", "plot2.png"))

}
