% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embed_iframe.R
\name{embed_slide}
\alias{embed_slide}
\title{embed slide by iframe}
\usage{
embed_slide(url, width = "100\%", height)
}
\arguments{
\item{url}{web-slide url. Slideshare, SpeakerDeck, and html slides are now available.}

\item{width}{ignored.}

\item{height}{ignored.}
}
\description{
embed slide by iframe
}
\examples{
# SlideShare url
embed_slide("http://www.slideshare.net/makotohirakawa3/osakastan2-chap51")
# SpeakerDeck url
embed_slide("https://speakerdeck.com/yamano357/tokyor49-stringr-stringi")
# other html-slide url
embed_slide("http://kz-md.net/stat/tmp_box/intro_tubeplayR.html#/")

}

