% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_prj.R
\name{set_prj}
\alias{set_prj}
\title{Set my R-Project.}
\usage{
set_prj(path = ".", index = FALSE)
}
\arguments{
\item{path}{Path to Project.}

\item{index}{Logical. If `true` then make "index.Rmd" file. This file is quite plane R Markdown file.}
}
\value{
Make "R", "docs", "data" directories in this Project directory.
}
\description{
Set my R-Project.
}
\examples{
set_prj()
}
