% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zousan_document.R
\name{zousan_document}
\alias{zousan_document}
\title{Convert to a html document for zousan.}
\usage{
zousan_document(fig_width = 7, fig_height = 5, fig_retina = 2,
  fig_caption = TRUE, self_contained = FALSE, dev = "png",
  keep_md = FALSE, mathjax = "default",
  md_extensions = "-ascii_identifiers", template = "default",
  katex = FALSE, lightbox = TRUE, showhide = TRUE, bs_panel = TRUE,
  bs_label = TRUE, ...)
}
\arguments{
\item{fig_width}{Default width (in inches) for figures}

\item{fig_height}{Default width (in inches) for figures}

\item{fig_retina}{Scaling to perform for retina displays (defaults to 2, which
currently works for all widely used retina displays). Set to \code{NULL} to
prevent retina scaling. Note that this will always be \code{NULL} when
\code{keep_md} is specified (this is because \code{fig_retina} relies on
outputting HTML directly into the markdown document).}

\item{fig_caption}{\code{TRUE} to render figures with captions}

\item{self_contained}{Produce a standalone HTML file with no external
dependencies, using data: URIs to incorporate the contents of linked
scripts, stylesheets, images, and videos. Note that even for self contained
documents MathJax is still loaded externally (this is necessary because of
it's size).}

\item{dev}{Graphics device to use for figure output (defaults to png)}

\item{keep_md}{Keep the markdown file generated by knitting.}

\item{mathjax}{Include mathjax. The "default" option uses an https URL from a
MathJax CDN. The "local" option uses a local version of MathJax (which is
copied into the output directory). You can pass an alternate URL or pass
\code{NULL} to exclude MathJax entirely.}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link{rmarkdown_format}} for
additional details.}

\item{template}{Pandoc template to use for rendering. Pass "default" to use
the rmarkdown package default template; pass \code{NULL} to use pandoc's
built-in template; pass a path to use a custom template that you've created.
Note that if you don't use the "default" template then some features of
\code{html_document} won't be available (see the Templates section below for
more details).}

\item{katex}{Logical. If TRUE, then use katex instead of mathjax.}

\item{lightbox}{Logical. If TRUE, then set lightbox automatically.}

\item{showhide}{Logical. If TRUE, you could set show/hide button.}

\item{bs_panel}{Logical. If TRUE, you could use Bootstrap panel.}

\item{bs_label}{Logical. if TRUE, you could use Bootstrap label.}

\item{...}{Additional function arguments to pass to the
base R Markdown HTML output formatter \code{\link{html_document_base}}}
}
\description{
html document including some tools.
}
