% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_rmd_ja.R
\name{new_rmd_ja}
\alias{new_rmd_ja}
\title{Make new .Rmd file for Japanese.}
\usage{
new_rmd_ja(file, path = ".", type = "html", systime = TRUE, title = "",
  subtitle = "", author = "")
}
\arguments{
\item{path}{Path to directory for output.}

\item{type}{Charactor or charactor vector. If `type = "html"` then `output: html_document`. If `type = "revealjs` then `output: revealjs::revealjs_presentation`.}

\item{systime}{Logical. If `TRUE` then set `sys.time` to `date: `.}

\item{title}{Charactor. If ""(default) then set file name.}

\item{subtitle}{Charactor. set `subtitle: `.}

\item{author}{Charactor. set `author: `.}

\item{Charactor.}{File output R Markdown file name.}
}
\description{
Make new .Rmd file for Japanese.
}

